refactor: move example project to submodule
